refactor: extract useInfiniteScroll hook from Gallery for reusability
